requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
aiohttp==3.9.1
aiosqlite==0.19.0
aiofiles==23.2.1
aiocron==1.8
loguru==0.7.2
pydantic==2.5.2
typing-extensions==4.9.0
cryptography==41.0.5
asyncio==3.4.3
async-timeout==4.0.3
pytz==2023.3
cachetools==5.3.2
setuptools==69.0.2
wheel==0.41.3
colorama==0.4.6
httpx==0.25.2
psutil==5.9.6
pytest==8.0.0
pytest-asyncio==0.23.5
pyjwt==2.8.0
pandas==2.2.3
openpyxl==3.1.5
numpy==1.26.4
python-dateutil==2.9.0.post0
apscheduler==3.11.0
tzlocal==5.3 